LPS-158029 Create Build Action Hook
