+++ modules/apps/site-initializer/site-initializer-testray/extra/remote-app/src/pages/Project/Routines/useRoutineActions.ts
@@ -13,23 +13,15 @@
  */
 
 import {useMutation} from '@apollo/client';
-import {useNavigate} from 'react-router-dom';
 
-import {
-	DeleteBuild,
-	DeleteRoutine,
-	UpdateBuild,
-} from '../../../graphql/mutations';
-import {TestrayBuild, TestrayRoutine} from '../../../graphql/queries';
+import {DeleteRoutine} from '../../../graphql/mutations';
+import {TestrayRoutine} from '../../../graphql/queries';
 import useFormModal from '../../../hooks/useFormModal';
 import i18n from '../../../i18n';
 
 const useRoutineActions = () => {
 	const formModal = useFormModal();
-	const [onUpdateBuild] = useMutation(UpdateBuild);
-	const [onDeleteBuild] = useMutation(DeleteBuild);
 	const [onDeleteRoutine] = useMutation(DeleteRoutine);
-	const navigate = useNavigate();
 
 	const modal = formModal.modal;
 
@@ -46,63 +38,6 @@ const useRoutineActions = () => {
 			{
 				action: ({id}: TestrayRoutine) =>
 					onDeleteRoutine({variables: {id}})
-						.then(() => modal.onSave())
-						.catch(modal.onError),
-				name: i18n.translate('delete'),
-			},
-		],
-		actionsBuild: [
-			{
-				action: () => alert('Archive'),
-				name: i18n.translate('archive'),
-			},
-			{
-				action: (testrayBuild: TestrayBuild) =>
-					navigate(`build/${testrayBuild.id}/update`),
-				name: i18n.translate('edit'),
-			},
-			{
-				action: (testrayBuild: TestrayBuild) =>
-					onUpdateBuild({
-						update(cache, {data: {updateBuild}}) {
-							cache.modify({
-								fields: {
-									builds(buildCache) {
-										return {
-											...buildCache,
-											items: buildCache.items.map(
-												(build: TestrayBuild) => {
-													if (
-														build.id ===
-														testrayBuild.id
-													) {
-														return {
-															...build,
-															promoted:
-																updateBuild.promoted,
-														};
-													}
-
-													return build;
-												}
-											),
-										};
-									},
-								},
-							});
-						},
-						variables: {
-							data: {
-								promoted: !testrayBuild.promoted,
-							},
-							id: testrayBuild.id,
-						},
-					}).then(() => modal.onSave(null, {forceRefetch: false})),
-				name: i18n.translate('promote'),
-			},
-			{
-				action: ({id}: any) =>
-					onDeleteBuild({variables: {id}})
 						.then(() => modal.onSave())
 						.catch(modal.onError),
 				name: i18n.translate('delete'),
